+++ processor/src/test/resources/input/com/example/default_props/PropDefaultWithColorfulName.java
@@ -1,0 +1,25 @@
+package com.example.default_props;
+
+import react4j.annotations.Prop;
+import react4j.annotations.PropDefault;
+import react4j.annotations.ReactComponent;
+import react4j.core.BaseState;
+import react4j.core.Component;
+import react4j.core.ReactNode;
+
+@ReactComponent
+abstract class PropDefaultWithColorfulName
+  extends Component<BaseState>
+{
+  @PropDefault
+  static final String DEFAULT_MY_PROP12$23 = "Foo";
+
+  @Prop
+  protected abstract String getMyProp12$23();
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}